toggleWishItems replaced to thunk actions
